+++ Tests/Tests.swift
@@ -5,15 +5,15 @@ import XCTest
 class Tests: XCTestCase {
 	
 	func testSuspendingRouteHandlingInInitializers() {
-		let routeController = RouteController(rootViewController: UIViewController())
+		let routeController = RouteController(root: UIViewController())
 		XCTAssertTrue(routeController.isPreparedForRouting)
 		
-		let otherController = RouteController(rootViewController: UIViewController(), ready: false)
+		let otherController = RouteController(root: UIViewController(), ready: false)
 		XCTAssertFalse(otherController.isPreparedForRouting)
 	}
 	
 	func testSuspendingRouteHandlingTemporally() {
-		let routeController = RouteController(rootViewController: UIViewController(), ready: true)
+		let routeController = RouteController(root: UIViewController(), ready: true)
 		XCTAssertTrue(routeController.isPreparedForRouting)
 		
 		routeController.suspendHandlingRoutes()
@@ -30,8 +30,8 @@ class Tests: XCTestCase {
 	}
 	
 	func testClearOfStoredRoutes() {
-		let routeController = RouteController(rootViewController: UIViewController(), ready: false)
-		let testRoute = Route(identifier: RouteIdentifier(rawValue: "test")) { _, _ in return true }
+		let routeController = RouteController(root: UIViewController(), ready: false)
+		let testRoute = Route<UIViewController>(identifier: RouteIdentifier(rawValue: "test")) { _, _ in return true }
 		routeController.open(testRoute)
 		XCTAssertNotNil(routeController.storedRoute)
 		
@@ -44,8 +44,8 @@ class Tests: XCTestCase {
 	
 	func testAutomaticExecutionOfStoredRoutes() {
 		let exp = expectation(description: "routeExecution")
-		let routeController = RouteController(rootViewController: UIViewController(), ready: false)
-		let testRoute = Route(identifier: RouteIdentifier(rawValue: "test")) { _, _ in
+		let routeController = RouteController(root: UIViewController(), ready: false)
+		let testRoute = Route<UIViewController>(identifier: RouteIdentifier(rawValue: "test")) { _, _ in
 			exp.fulfill()
 			return true
 		}
@@ -60,19 +60,19 @@ class Tests: XCTestCase {
 	}
 	
 	func testAutomaticOpeningOfRouteConvertibles() {
-		let routeController = RouteController(rootViewController: UIViewController(), ready: true)
+		let routeController = RouteController(root: UIViewController(), ready: true)
 		XCTAssertTrue(routeController.open(SomeRouteConvertible().route))
 	}
 	
 	func testHandlingOfRouteConvertibleFailure() {
-		let routeController = RouteController(rootViewController: UIViewController(), ready: true)
+		let routeController = RouteController(root: UIViewController(), ready: true)
 		XCTAssertFalse(routeController.open(FailingRouteConvertible().route))
 	}
 	
 	func testRouteActionableDestination() {
 		let exp = expectation(description: "routeExecution")
-		let routeController = RouteController(rootViewController: UIViewController())
-		let testRoute = Route(identifier: RouteIdentifier(rawValue: "test")) { _, _ in
+		let routeController = RouteController(root: UIViewController())
+		let testRoute = Route<UIViewController>(identifier: RouteIdentifier(rawValue: "test")) { _, _ in
 			
 			let actionable = ActionableObject()
             actionable.setRouteAction {
@@ -116,18 +116,18 @@ class Tests: XCTestCase {
     }
 	
 	func testNestedAnyRouteInitializers() {
-		let testRoute = Route(identifier: RouteIdentifier(rawValue: "test")) { _, _ in return true }
+		let testRoute = Route<UIViewController>(identifier: RouteIdentifier(rawValue: "test")) { _, _ in return true }
 		let nestedRoute = Route(route: testRoute)
 		
 		XCTAssert(testRoute.isSuspendable == nestedRoute.isSuspendable)
 		
-		let routeA = testRoute.route(fromRootViewController: UIViewController(), options: nil)
-		let routeB = nestedRoute.route(fromRootViewController: UIViewController(), options: nil)
+		let routeA = testRoute.route(fromRoot: UIViewController(), options: nil)
+		let routeB = nestedRoute.route(fromRoot: UIViewController(), options: nil)
 		XCTAssert(routeA == routeB)
 	}
 	
 	func testNestedAnyRouteInitializerOverrideSuspendable() {
-		let testRoute = Route(identifier: RouteIdentifier(rawValue: "test")) { _, _ in return true }
+		let testRoute = Route<UIViewController>(identifier: RouteIdentifier(rawValue: "test")) { _, _ in return true }
 		let nestedRoute = Route(route: testRoute, isSuspendable: false)
 		
 		XCTAssert(testRoute.isSuspendable != nestedRoute.isSuspendable)
@@ -161,7 +161,7 @@ class Tests: XCTestCase {
     func testRoutablePatternMatching() {
         let exp = expectation(description: "pattern match")
         let id = RouteIdentifier(rawValue: "test")
-        let route = Route(identifier: id) { _, _ in return true }
+        let route = Route<UIViewController>(identifier: id) { _, _ in return true }
         
         switch route {
         case id: exp.fulfill()